refactor: Remove importing directly from `utils.js` in `world/*`
