+++ src/world/tree_generation.js
@@ -1,5 +1,5 @@
 import { alea } from "../alea/alea.js";
-import { rangeFrom, rangeList } from "../utils.js";
+import { rangeFrom, rangeList } from "../utils/array_utils.js";
 import { BaseGenerator } from "./base_generator.js";
